core/vm: consolidate cas20 abi helpers
