+++ core/vm/cas20_factory.go
@@ -311,53 +311,3 @@ func encodeCAS20CreatedData(c cas20CreateParams) []byte {
 		abiBytes(variantParams),
 	)
 }
-
-func readBytesArray(args []byte, argIndex int) ([][]byte, error) {
-	L := uint64(len(args))
-	base, ok := wordU64(args, uint64(argIndex)*32)
-	if !ok || base > L || L-base < 32 {
-		return nil, ErrExecutionReverted
-	}
-	n, ok2 := wordU64(args, base)
-	if !ok2 {
-		return nil, ErrExecutionReverted
-	}
-	arrData := base + 32
-	if n > (L-arrData)/32 {
-		return nil, ErrExecutionReverted
-	}
-	out := make([][]byte, n)
-	for i := uint64(0); i < n; i++ {
-		elemOff, ok2 := wordU64(args, arrData+i*32)
-		if !ok2 {
-			return nil, ErrExecutionReverted
-		}
-		pos := arrData + elemOff
-		if elemOff > L-arrData || pos > L || L-pos < 32 {
-			return nil, ErrExecutionReverted
-		}
-		elemLen, ok3 := wordU64(args, pos)
-		if !ok3 {
-			return nil, ErrExecutionReverted
-		}
-		start := pos + 32
-		if elemLen > L-start {
-			return nil, ErrExecutionReverted
-		}
-		out[i] = args[start : start+elemLen]
-	}
-	return out, nil
-}
-
-// An offset or length with dirty high bits is a malformed encoding, not a large number.
-func wordU64(args []byte, pos uint64) (uint64, bool) {
-	if pos > uint64(len(args)) || uint64(len(args))-pos < 32 {
-		return 0, false
-	}
-	for _, b := range args[pos : pos+24] {
-		if b != 0 {
-			return 0, false
-		}
-	}
-	return new(uint256.Int).SetBytes(args[pos+24 : pos+32]).Uint64(), true
-}
