+++ core/vm/cas20_token.go
@@ -276,146 +276,3 @@ func (t cas20Token) emit(topic0 common.Hash, a, b common.Address, value *uint256
 	v := value.Bytes32()
 	return t.ctx.AddLog([]common.Hash{topic0, addrKey(a), addrKey(b)}, v[:])
 }
-
-// --- ABI helpers ------------------------------------------------------------
-
-func readWord(args []byte, i int) (common.Hash, error) {
-	off := i * 32
-	if len(args) < off+32 {
-		return common.Hash{}, ErrExecutionReverted
-	}
-	return common.BytesToHash(args[off : off+32]), nil
-}
-
-func readAddress(args []byte, i int) (common.Address, error) {
-	w, err := readWord(args, i)
-	if err != nil {
-		return common.Address{}, err
-	}
-	a, ok := addressFromWord(w)
-	if !ok {
-		return common.Address{}, ErrExecutionReverted
-	}
-	return a, nil
-}
-
-// wordFitsIn is the check Solidity's external decoder makes for every type
-// narrower than a word: dirty high bytes are a malformed encoding, not a value.
-func wordFitsIn(w common.Hash, n int) bool {
-	for _, b := range w[:32-n] {
-		if b != 0 {
-			return false
-		}
-	}
-	return true
-}
-
-func u64FromWord(w common.Hash) (uint64, bool) {
-	if !wordFitsIn(w, 8) {
-		return 0, false
-	}
-	return new(uint256.Int).SetBytes(w.Bytes()).Uint64(), true
-}
-
-func addressFromWord(w common.Hash) (common.Address, bool) {
-	if !wordFitsIn(w, 20) {
-		return common.Address{}, false
-	}
-	return common.BytesToAddress(w.Bytes()), true
-}
-
-func readU64(args []byte, i int) (uint64, error) {
-	w, err := readWord(args, i)
-	if err != nil {
-		return 0, err
-	}
-	v, ok := u64FromWord(w)
-	if !ok {
-		return 0, ErrExecutionReverted
-	}
-	return v, nil
-}
-
-func readU256(args []byte, i int) (*uint256.Int, error) {
-	w, err := readWord(args, i)
-	if err != nil {
-		return nil, err
-	}
-	return new(uint256.Int).SetBytes(w.Bytes()), nil
-}
-
-func encU256(v *uint256.Int) []byte {
-	b := v.Bytes32()
-	return b[:]
-}
-
-func encBool(b bool) []byte {
-	out := make([]byte, 32)
-	if b {
-		out[31] = 1
-	}
-	return out
-}
-
-func encString(s string) []byte { return encodeTuple(abiString(s)) }
-
-// --- ABI encoding primitives ------------------------------------------------
-
-type abiPart struct {
-	word    common.Hash
-	dynamic bool
-	tail    []byte
-}
-
-func abiWord(w common.Hash) abiPart { return abiPart{word: w} }
-
-func abiBytes(b []byte) abiPart {
-	padded := (len(b) + 31) / 32 * 32
-	tail := make([]byte, 32+padded)
-	l := uint256.NewInt(uint64(len(b))).Bytes32()
-	copy(tail[:32], l[:])
-	copy(tail[32:], b)
-	return abiPart{dynamic: true, tail: tail}
-}
-
-func abiString(s string) abiPart { return abiBytes([]byte(s)) }
-
-func abiWordArray(words []common.Hash) abiPart {
-	tail := make([]byte, 0, 32*(len(words)+1))
-	l := uint256.NewInt(uint64(len(words))).Bytes32()
-	tail = append(tail, l[:]...)
-	for _, w := range words {
-		tail = append(tail, w[:]...)
-	}
-	return abiPart{dynamic: true, tail: tail}
-}
-
-func encodeTuple(parts ...abiPart) []byte {
-	head := make([]byte, 0, 32*len(parts))
-	tail := make([]byte, 0)
-	tailStart := uint64(32 * len(parts))
-	for _, p := range parts {
-		if !p.dynamic {
-			head = append(head, p.word[:]...)
-			continue
-		}
-		off := uint256.NewInt(tailStart + uint64(len(tail))).Bytes32()
-		head = append(head, off[:]...)
-		tail = append(tail, p.tail...)
-	}
-	return append(head, tail...)
-}
-
-// abi.encode of one dynamic struct wraps it in a one-element tuple, so the
-// result opens with an offset word (0x20) before the struct's own head/tail.
-func abiEncodeStruct(members ...abiPart) []byte {
-	return encodeTuple(abiPart{dynamic: true, tail: encodeTuple(members...)})
-}
-
-func readBytesArg(args []byte, argIndex int) ([]byte, error) {
-	s, err := readStringArg(args, argIndex)
-	if err != nil {
-		return nil, err
-	}
-	return []byte(s), nil
-}
